updating properties on batch updates
Lesson learned from replication: if a single
feature is selected and there are > 1
operations affected by the batch the
property panel will receive all ops and
show M/V (multiple values) for many properties.
To avoid this, we dispatch an update only the selected features.
This problem will never occour if running
in single-user mode.